pyproject.toml
[tool.poetry]
name = "clothes-magazine-project"
version = "0.1.0"
discription = "Poetry environnement config"
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.7"
psycopg2 = "^2.9.9"
django-environ = "^0.11.2"
makefile = "^1.1.0"
docker = "^7.1.0"
django-ninja = "^1.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
